split services into services and entities
