+++ welcome/welcome.go
@@ -11,8 +11,8 @@ import (
 	"github.com/charmbracelet/bubbles/textinput"
 	tea "github.com/charmbracelet/bubbletea"
 	"github.com/charmbracelet/lipgloss"
-	"github.com/onioncall/cli-squa/cli/create"
-	"github.com/onioncall/cli-squa/cli/join"
+	"github.com/onioncall/squa/create"
+	"github.com/onioncall/squa/join"
 )
 
 var (
